Music Test
Added a music test for the game to test if the sound sync is working properly. There are two songs, which can be swapped between by switching the songNum in config. Also ended up adding in hitsounds (which can be toggled in config), because I needed it for testing sync. Made songs a bit less hardcoded as well, so additional songs are easier to add in the future..