Implemented third song
This third song takes the place of the original placeholder song as the tutorial. It's not currently easy enough to be an actual tutorial, but the beginning is easy enough. There is currently an issue with the game having a mind of its own and throwing the offset of the song off, making it out of sync.